+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-- fix: VPC used in example
+- feature: VPC used in example
+- Feature: Use a CMK in the complete example
 
 ## [1.0.3] - 2022-06-24
 ### Description